refactor(book): restructure into hexagonal architecture layers
- Reorganize book module into api, application, domain, and infrastructure
- Move DTOs to api/ as inbound port contracts
- Move services to application/ layer
- Move entities, mappers, and repositories to infrastructure/
- Relocate external API types (GoogleBooksResponse) to infrastructure/external
- Extract controllers to dedicated web layer (com.penrose.bibby.web)
- Convert ShelfFacade from empty class to interface with getShelfEntityById

Package structure now follows ports-and-adapters pattern:
  book/
    api/          - DTOs, facades (inbound ports)
    application/  - services (use cases)
    domain/       - Book, BookFactory (core business logic)
    infrastructure/
      entity/     - JPA entities
      external/   - Google Books API types
      mapping/    - entity-domain mappers
      repository/ - data access
