+++ src/main/java/com/penrose/bibby/cli/BookCommands.java
@@ -3,11 +3,11 @@ package com.penrose.bibby.cli;
 import com.penrose.bibby.library.author.Author;
 import com.penrose.bibby.library.author.AuthorEntity;
 import com.penrose.bibby.library.author.AuthorService;
-import com.penrose.bibby.library.book.domain.BookEntity;
-import com.penrose.bibby.library.book.domain.GoogleBooksResponse;
-import com.penrose.bibby.library.book.dto.BookRequestDTO;
-import com.penrose.bibby.library.book.mapping.BookMapper;
-import com.penrose.bibby.library.book.service.BookService;
+import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
+import com.penrose.bibby.library.book.infrastructure.external.GoogleBooksResponse;
+import com.penrose.bibby.library.book.api.BookRequestDTO;
+import com.penrose.bibby.library.book.infrastructure.mapping.BookMapper;
+import com.penrose.bibby.library.book.application.BookService;
 import com.penrose.bibby.library.bookcase.persistence.BookcaseEntity;
 import com.penrose.bibby.library.bookcase.service.BookcaseService;
 import com.penrose.bibby.library.shelf.*;
