+++ src/main/java/com/penrose/bibby/cli/BookcaseCommands.java
@@ -1,9 +1,9 @@
 package com.penrose.bibby.cli;
 
-import com.penrose.bibby.library.book.dto.BookDetailView;
-import com.penrose.bibby.library.book.domain.BookEntity;
-import com.penrose.bibby.library.book.service.BookService;
-import com.penrose.bibby.library.book.dto.BookSummary;
+import com.penrose.bibby.library.book.api.BookDetailView;
+import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
+import com.penrose.bibby.library.book.application.BookService;
+import com.penrose.bibby.library.book.api.BookSummary;
 import com.penrose.bibby.library.bookcase.persistence.BookcaseEntity;
 import com.penrose.bibby.library.bookcase.service.BookcaseService;
 import com.penrose.bibby.library.shelf.ShelfEntity;
