--- src/main/java/com/penrose/bibby/library/author/AuthorController.java
@@ -1,4 +1,0 @@
-package com.penrose.bibby.library.author;
-
-public class AuthorController {
-}
