--- src/main/java/com/penrose/bibby/library/book/BookFacade.java
@@ -1,4 +1,0 @@
-package com.penrose.bibby.library.book;
-
-public class BookFacade {
-}
