+++ src/main/java/com/penrose/bibby/library/book/api/BookFacade.java
@@ -1,0 +1,4 @@
+package com.penrose.bibby.library.book.api;
+
+public class BookFacade {
+}
