+++ src/main/java/com/penrose/bibby/library/book/application/BookService.java
@@ -1,17 +1,18 @@
-package com.penrose.bibby.library.book.service;
+package com.penrose.bibby.library.book.application;
 
 import com.penrose.bibby.library.author.Author;
 import com.penrose.bibby.library.author.AuthorEntity;
 import com.penrose.bibby.library.author.AuthorService;
 import com.penrose.bibby.library.book.domain.Book;
-import com.penrose.bibby.library.book.domain.BookEntity;
+import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
 import com.penrose.bibby.library.book.domain.BookFactory;
-import com.penrose.bibby.library.book.dto.BookDetailView;
-import com.penrose.bibby.library.book.dto.BookRequestDTO;
-import com.penrose.bibby.library.book.dto.BookSummary;
-import com.penrose.bibby.library.book.mapping.BookMapper;
-import com.penrose.bibby.library.book.repository.BookRepository;
+import com.penrose.bibby.library.book.api.BookDetailView;
+import com.penrose.bibby.library.book.api.BookRequestDTO;
+import com.penrose.bibby.library.book.api.BookSummary;
+import com.penrose.bibby.library.book.infrastructure.mapping.BookMapper;
+import com.penrose.bibby.library.book.infrastructure.repository.BookRepository;
 import com.penrose.bibby.library.shelf.ShelfEntity;
+import com.penrose.bibby.library.shelf.ShelfFacade;
 import com.penrose.bibby.library.shelf.ShelfService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
