--- src/main/java/com/penrose/bibby/library/book/domain/GoogleBookItems.java
@@ -1,4 +1,0 @@
-package com.penrose.bibby.library.book.domain;
-
-public record GoogleBookItems(VolumeInfo volumeInfo) {
-}
