--- src/main/java/com/penrose/bibby/library/book/dto/BookImportRequest.java
@@ -1,4 +1,0 @@
-package com.penrose.bibby.library.book.dto;
-
-public record BookImportRequest(String isbn) {
-}
