+++ src/main/java/com/penrose/bibby/library/book/infrastructure/external/BookImportRequest.java
@@ -1,0 +1,4 @@
+package com.penrose.bibby.library.book.infrastructure.external;
+
+public record BookImportRequest(String isbn) {
+}
