+++ src/main/java/com/penrose/bibby/library/book/infrastructure/external/BookImportResponse.java
@@ -1,4 +1,4 @@
-package com.penrose.bibby.library.book.dto;
+package com.penrose.bibby.library.book.infrastructure.external;
 
 import java.util.List;
