+++ src/main/java/com/penrose/bibby/library/book/infrastructure/external/GoogleBooksResponse.java
@@ -1,4 +1,4 @@
-package com.penrose.bibby.library.book.domain;
+package com.penrose.bibby.library.book.infrastructure.external;
 
 import java.util.List;
