+++ src/main/java/com/penrose/bibby/library/book/infrastructure/mapping/BookMapper.java
@@ -1,11 +1,11 @@
-package com.penrose.bibby.library.book.mapping;
+package com.penrose.bibby.library.book.infrastructure.mapping;
 
 import com.penrose.bibby.library.author.Author;
 import com.penrose.bibby.library.author.AuthorEntity;
 import com.penrose.bibby.library.author.AuthorMapper;
 import com.penrose.bibby.library.book.domain.AvailabilityStatus;
 import com.penrose.bibby.library.book.domain.Book;
-import com.penrose.bibby.library.book.domain.BookEntity;
+import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
 import com.penrose.bibby.library.shelf.Shelf;
 import com.penrose.bibby.library.shelf.ShelfDomainRepositoryImpl;
 import com.penrose.bibby.library.shelf.ShelfEntity;
