+++ src/main/java/com/penrose/bibby/library/book/infrastructure/repository/BookRepository.java
@@ -1,8 +1,8 @@
-package com.penrose.bibby.library.book.repository;
+package com.penrose.bibby.library.book.infrastructure.repository;
 
-import com.penrose.bibby.library.book.domain.BookEntity;
-import com.penrose.bibby.library.book.dto.BookDetailView;
-import com.penrose.bibby.library.book.dto.BookSummary;
+import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
+import com.penrose.bibby.library.book.api.BookDetailView;
+import com.penrose.bibby.library.book.api.BookSummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
