+++ src/main/java/com/penrose/bibby/library/shelf/Shelf.java
@@ -1,7 +1,7 @@
 package com.penrose.bibby.library.shelf;
 
 import com.penrose.bibby.library.book.domain.Book;
-import com.penrose.bibby.library.book.domain.BookEntity;
+import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
 import com.penrose.bibby.library.bookcase.domain.Bookcase;
 
 import java.util.List;
