+++ src/main/java/com/penrose/bibby/library/shelf/ShelfDomainRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.penrose.bibby.library.shelf;
 
 import com.penrose.bibby.library.book.domain.Book;
-import com.penrose.bibby.library.book.repository.BookDomainRepository;
+import com.penrose.bibby.library.book.infrastructure.repository.BookDomainRepository;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
