+++ src/main/java/com/penrose/bibby/library/shelf/ShelfFacade.java
@@ -1,4 +1,8 @@
 package com.penrose.bibby.library.shelf;
 
-public class ShelfFacade {
+import java.util.Optional;
+
+public interface ShelfFacade {
+
+    Optional<ShelfEntity> getShelfEntityById(Long shelfId);
 }
