+++ src/main/java/com/penrose/bibby/web/author/AuthorController.java
@@ -1,0 +1,4 @@
+package com.penrose.bibby.web.author;
+
+public class AuthorController {
+}
